feat(platform): 完成配额计费闭环与 Agent 安全加固
- 打通 SKU 定价、配额冻结与结算、权益发放与撤销，补齐 Gateway、BFF、Member、Pay 和 Group 全链路
- 强化内部令牌、身份归属、提示词与日志安全，完善 Agent SSE 生命周期、执行账本、检查点及长期记忆隔离
- 补齐一键全栈启动、security/E2E/Agent SSE/memory smoke、离线评测与并发压测脚本
- 增加跨模块单元与集成测试及运维文档，统一本地开发配置和安全默认值

+++ s-pay-mall-ddd-market/s-pay-mall-ddd-domain/src/main/java/com/aigroup/paymall/domain/benefit/service/IBenefitEventService.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface IBenefitEventService {
 
-    void publishGroupBuyCompletedEvents(List<String> orderIds, Integer bonusQuota);
+    void publishGroupBuyCompletedEvents(List<String> orderIds, Long bonusQuota);
 
     void publishGroupBuyRevokedEvents(List<String> orderIds);
